Fix wrong SSL 2.0 version and remove SSL 1.0 (bug in older JA4 spec)
Fix version for SSL 2.0 and remove version for SSL 1.0, porting a bug
fix [1] made in the JA4 spec in August 2024.  While here, also fix some
versions mentioned in strings to match the code.

SSL 2.0 [2][3] uses a version field of 0x0002, not 0x0200.  SSL 1 never
existed outside of Netscape, as the original design was iterated upon to
become SSL 2 before the first public version of SSL.  I don't think it's
public knowledge what the version field for SSL 1.0 looked like, or if
it even was two bytes large or at the same offset on the wire.

Version field 0x0100, that nfdump is currently misattributing to SSL
1.0, was used by an early pre-RFC4347 implementation of DTLS in OpenSSL
before 0.9.8f [2], when OpenSSL switched to the version field specified
by RFC4347.

[1] https://github.com/FoxIO-LLC/ja4/pull/150
[2] https://www-archive.mozilla.org/projects/security/pki/nss/ssl/draft02.html
[3] https://datatracker.ietf.org/doc/html/draft-hickman-netscape-ssl-00
[4] openssl/openssl@OpenSSL_0_9_8e...OpenSSL_0_9_8f